上传了组员个人草稿
